requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
psycopg2==2.9.3
python-decouple==3.6
python-slugify==5.0.2
pytz==2021.3
sqlparse==0.4.2
text-unidecode==1.3
whitenoise==6.1.0
